refactor: GraphQL client config option 분리
