+++ web/src/graphql/config/index.ts
@@ -1,0 +1,2 @@
+export { default as cacheConfig } from './cacheConfig'
+export { default as typeDefs } from './typeDefs'
